Add files via upload
In this talk, we looked in more detail at the diagrammatic composition of BCFW recursions for tree-level, gluonic Scattering Amplitudes of both arbitrary particle number and helicity state.  We then briefly looked at scalar QED to interpret the results of BCFW recursion when the seed amplitudes are insufficient to fix all on-shell, gauge-invariant information.